requirements-color-detection.txt
numpy==1.23.4
Pillow==9.2.0
